pyproject.toml
[project]
name = "drive-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.191.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
    "ty>=0.0.20",
]
